:recycle: Refactor: Shared 페이지 컴포넌트 모듈화
