+++ src/components/Shared/SharedCards/Card/Card.module.scss
