+++ src/components/Shared/SharedCards/SharedCards.module.scss
